requirements.txt
tqdm==4.58.0
numpy==1.19.2
ImageHash==4.2.0
h5py==3.2.1
lmdb==1.1.1
Pillow==8.1.2
